# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PHP Tinder Backend API",
        "description": "Technical assignment backend API built with Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "List recommended people",
                "operationId": "48f7032d1333ee6a10b694ad02ce52ef",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "Current user ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Missing X-User-Id header"
                    }
                }
            }
        },
        "/api/people/{id}/like": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Like a person",
                "operationId": "f948dd4d5c3dd79ae7ad0e405b12a500",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "Current user ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID to like",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person liked"
                    },
                    "400": {
                        "description": "Validation error (missing header or like yourself)"
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/people/{id}/dislike": {
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Dislike a person",
                "operationId": "0c1b052305229a3750fded98090476c0",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "Current user ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Person ID to dislike",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person disliked"
                    },
                    "400": {
                        "description": "Validation error (missing header or dislike yourself)"
                    },
                    "404": {
                        "description": "Person not found"
                    }
                }
            }
        },
        "/api/people/liked": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "List liked people",
                "operationId": "eab4da95238438a87098915b90ee9a87",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "Current user ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of liked people"
                    },
                    "400": {
                        "description": "Missing X-User-Id header"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "People",
            "description": "People"
        }
    ]
}